build: Disable parallel build (make -j) for now.
The make manual to not recommend using parallel build when archives
are used. Curently sometimes when building in parallel, linking stage
is reached before all dependencies are built and linking fails.

Thats why parallel build is disabled for now, until it is proven to
be working correctly in all cases.